dev-libs/msgpuck: drop 1.1.3
dev-libs/msgpuck-1.1.3 and dev-libs/msgpuck-2.0 (or -9999) have
different SLOTs, so it is expected that they can be installed at the
same time on the same system. However it is not possible due to file
collisions:

* /usr/include/msgpuck.h
* /usr/lib64/libmsgpuck.a

Fixing of this would complicate building of dependent projects (setup
custom include and library paths), so I decided to drop the old version
instead.