dev-libs/msgpuck: use tarantool's fork consistently
Before this commit the upstream was used for releases, but tarantool's
fork for the last git version. Since this overlay is about tarantool, it
seems logical to provide tarantool's fork here.

The upstream and tarantool's fork are identical for most applications
(at least now).

Aside of this, the library is not more header-only. Updated the
description by copying it from the github.com page.