dev-db/tarantool: drop gcc/clang version checks
gcc-4.5 and clang-3.2 are very old versions: the main Gentoo tree does
not provide them for a long time. So there is no sense to check and
worth to clean up this code.

Also I doubt that we need an explicit dependency on a compiler, so
removed it as well.