www-servers/nginx: drop the ebuild
This ebuild provides nginx with tarantool upstream module [1]. It mostly
just copy-paste from the ebuild from the main Gentoo tree.

Patching it each time when a next nginx version lands to the main tree
is cumbersome, so the ebuild was not updated for a quite long time. It
seems, the ebuild becomes too old now: the automated QA check reports
the following problem:

 | $ pmaint --config /var/lib/repo-mirror-ci/data/etc/portage regen --use-local-desc --pkg-desc-index -t 32 tarantool
 |  * www-servers/nginx-1.17.8: failed sourcing ebuild: perl-module: EAPI 6 not supported, (perl-module.eclass, line  32:  called die)     [FATAL]
 |  * Cache regen failed with 1

This commit just removes the obsolete version of the web server. I don't
want to support it until there will be a convenient mechanism like one
proposed in [2].

Sorry for possible inconvenience.

[1]: https://github.com/tarantool/nginx_upstream_module
[2]: https://bugs.gentoo.org/573710

Fixes https://bugs.gentoo.org/859109